vsp2driver: Add struct vsp2_pipeline num_video
vsp2driver clears the parameter of vspm, when stop_streaming is executed.
However stop_streaming is not executed,when 1 rpf and 1 wpf are used and
the result of s_stream() for rpf is error. Because num_video is deleted.
This commit adds struct vsp2_pipeline num_video to support the above error.

Signed-off-by: Sojiro Kusunoki <sojiro.kusunoki.zj@ps.hitachi-solutions.com>