vsp2driver: Fix vspm channel designation
This commit adds the channel designation for vspm.
The purposes of this commit is to use the fix channel
of vsp2 in vsp2driver.
By default, vsp2driver does not use the channel designation.
Therefore vsp2driver uses not the fix channel but the empty channel.
If you want to use the fix channel in vsp2driver, refer to the following
file.
  vsp2driver/document/renesas,vsp2.txt
Note the following
If you use the fix channel in one application,
you can not use the same channel in other application.

Signed-off-by: Sojiro Kusunoki <sojiro.kusunoki.zj@ps.hitachi-solutions.com>